generalize ClientRecipeBook
