+++ common/src/main/java/marsh/town/brb/brewingstand/BrewingClientRecipeBook.java
@@ -3,19 +3,11 @@ package marsh.town.brb.brewingstand;
 import marsh.town.brb.generic.GenericClientRecipeBook;
 import marsh.town.brb.loaders.PotionLoader;
 import marsh.town.brb.recipe.BRBRecipeBookCategories;
-import net.minecraft.stats.RecipeBook;
-import net.minecraft.world.inventory.RecipeBookType;
 
 import java.util.List;
 
-public class BrewingClientRecipeBook extends RecipeBook implements GenericClientRecipeBook {
-    private boolean filteringCraftable;
-
-    public boolean isFilteringCraftable() {
-        return filteringCraftable;
-    }
-
-    public List<BrewableResult> getResultsForCategory(BRBRecipeBookCategories category) {
+public class BrewingClientRecipeBook extends GenericClientRecipeBook {
+    public List<BrewableResult> getCollectionsForCategory(BRBRecipeBookCategories category) {
         List<BrewableResult> results = PotionLoader.POTIONS;
 
         if (category == BRBRecipeBookCategories.BREWING_SPLASH_POTION) {
@@ -25,13 +17,5 @@ public class BrewingClientRecipeBook extends RecipeBook implements GenericClient
         }
 
         return results;
-    }
-
-    public boolean isFiltering(RecipeBookType category) {
-        return filteringCraftable;
-    }
-
-    public void setFilteringCraftable(boolean filteringCraftable) {
-        this.filteringCraftable = filteringCraftable;
     }
 }
