+++ common/src/main/java/marsh/town/brb/generic/GenericClientRecipeBook.java
@@ -1,5 +1,20 @@
 package marsh.town.brb.generic;
 
-public interface GenericClientRecipeBook {
-    void setFilteringCraftable(boolean b);
+import net.minecraft.stats.RecipeBook;
+import net.minecraft.world.inventory.RecipeBookType;
+
+public abstract class GenericClientRecipeBook extends RecipeBook {
+    private boolean filteringCraftable;
+
+    public boolean isFilteringCraftable() {
+        return filteringCraftable;
+    }
+
+    public boolean isFiltering(RecipeBookType category) {
+        return filteringCraftable;
+    }
+
+    public void setFilteringCraftable(boolean filteringCraftable) {
+        this.filteringCraftable = filteringCraftable;
+    }
 }
