+++ common/src/main/java/marsh/town/brb/smithingtable/SmithingClientRecipeBook.java
@@ -5,21 +5,13 @@ import marsh.town.brb.recipe.BRBRecipeBookCategories;
 import marsh.town.brb.recipe.smithing.BRBSmithingTransformRecipe;
 import marsh.town.brb.recipe.smithing.BRBSmithingTrimRecipe;
 import net.minecraft.core.RegistryAccess;
-import net.minecraft.stats.RecipeBook;
-import net.minecraft.world.inventory.RecipeBookType;
 import net.minecraft.world.inventory.SmithingMenu;
 import net.minecraft.world.item.crafting.*;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class SmithingClientRecipeBook extends RecipeBook implements GenericClientRecipeBook {
-    private boolean filteringCraftable;
-
-    public boolean isFilteringCraftable() {
-        return filteringCraftable;
-    }
-
+public class SmithingClientRecipeBook extends GenericClientRecipeBook {
     public List<SmithingRecipeCollection> getCollectionsForCategory(BRBRecipeBookCategories group, SmithingMenu smithingScreenHandler, RegistryAccess registryAccess, RecipeManager recipeManager) {
         List<RecipeHolder<SmithingRecipe>> recipes = recipeManager.getAllRecipesFor(RecipeType.SMITHING);
         List<SmithingRecipeCollection> results = new ArrayList<>();
@@ -43,13 +35,5 @@ public class SmithingClientRecipeBook extends RecipeBook implements GenericClien
         }
 
         return results;
-    }
-
-    public boolean isFiltering(RecipeBookType category) {
-        return filteringCraftable;
-    }
-
-    public void setFilteringCraftable(boolean filteringCraftable) {
-        this.filteringCraftable = filteringCraftable;
     }
 }
